requirements.txt
numpy
tensorflow
keras
Pillow
cupy-cuda12x